+++ fiul-rest/fiul-rest-notification/pom.xml
@@ -15,10 +15,12 @@
         <dependency>
             <groupId>io.finarkein.fiu</groupId>
             <artifactId>fiul-notification-in-mem-pub</artifactId>
+            <scope>runtime</scope>
         </dependency>
         <dependency>
             <groupId>io.finarkein.fiu</groupId>
             <artifactId>fiul-notification-callbacks-in-mem-sub</artifactId>
+            <scope>runtime</scope>
         </dependency>
         <dependency>
             <groupId>io.finarkein.fiu</groupId>
